Small bug fixes in Login and Register pages. Started the Portfolio page.
There were a few bug in the Login and Register pages. Started a global CSS called 'main_style' to help clean things up a bit.

The Portfolio page has been started. A 'hamburger' style menu button has been added to the upper right which opens a menu bar along on the whole width of the page. This will hold the links to the Portfolio and Transaction pages, and will be shared among them.

Next will be to start adding long rectangular dynamic <div> boxes that will display a stock ticker and some information from the users portfolio. 

Time permitting, it would be nice to have a button on the left of each ticker info <div> that opened up a candlestick chart underneath for the past period and some extra information for the user.