Side bar working after some finger grease!
Side bar works mostly as expected. Only issue is on the initial load of the page were the link boxes click to the word instead of transitioning like they correctly do the rest of the time.

Next will be to add dynamic <div> boxes for each of the stocks to build up a dynamic portfolio viewer. The same <div> should be close to how the Transaction information <div> will work.

Back-end will be next. Until next time!