Sidebar fixed, and with new features!
The clicking instead of transitioning of side bar links the first time, now transitions right away.

The menu was moved from <nav> to <header>, just makes more sense. The <header> is now fixed, along with the side bar when it is active.

Now when you open the side bar, it dims the rest of the page. This is in hopes to help bring attention to the side bar and not from the rest of the page's distractions.

Next will be to build dynamic <div> boxes to show case each stock that is purchased.

Next will be to get the Transaction page  dynamic <div> are ready for the final step.

Final plans are to introduce PHP and PostgreSQL for back-end functionalities and really help bring the whole page together.