Back-end is in progress.
Back-end has been started, but has some ways to go. So far only the Register and Login page have been started, and only the Register page is working correctly with the database.

Next will be to finish the Register and Login pages, and to start the Portfolio and Transaction pages.

Finally, connecting with finance information and debugging will, hopefully, be looked at Sunday night, and finished Monday.