Login and Registration 90% complete. Also, page security implemented.
Login and Registration 90% complete. The last 10% will be to check if user is already registered, bring them to login, and if the user doesn't log in right, put up some warning on login page. Last will be to implement a password recovery.
Next will be to work on a dynamic Portfolio page that will display the stock information for each stock in the portfolio. This information will be pulled from the database with PHP, then loop through each stock, placing a <div> of information. Hopefully will be able to copy most of this over to the Transactions page with just a small change to the code and SQL queries. Note to self, don't forget to clean all database queries before sending them off!