Back-end for login is 99% done. Haven't gone any further with Portfolio.
Taking time to do things right matters. Back-end is set so if someone is logged in, that email cannot be used to login again. If you try to register and the email is already in use, you'll be told the email is already chosen. Users are stopped from looking at pages they are not authorized to look at, and will be redirected to the appropriate page. Errors are available when needed.

Only thing to do next is to start on the Portfolio by linking to financial data and start the transactions table in the database. The question is now, to continue adding functionality to the 'EZDB' class, or start making multiple classes.

A good 2-3 hours of code reformatting would be a good idea before getting to much further. May be able to structure some of the code better and cleaner. Also, could fine loopholes during the process. Cheers!