Login and Register pages are done. Started back-end for Portfolio page.
Login and Register pages are done. Just looking for bugs to fix, if any, and looking for security issues.

Portfolio and Transaction pages are starting to be linked to a 'transaction' table to start saving transactions. Transaction table schema works, but looking into changes it to better label each transaction.

Next will be to link access to stock data. Will be looking into IEX API. Will need to build a form above portfolio information to look up stock prices and buy and sell based on current prices.

If time permitting. It would be nice if the Portfolio page's stock information block would expand to show a chart underneath plotting the stocks history, and possible some other statistics such as MACD or even a regression for predictions based on previous trend using Tensorflow.js, by continuously training predicting as new data is created.