Most bugs fixed. Dropped 'transType' field. Stock buttons in works.
Most bug resulting from DB change have been fixed. Still an issue with the error messages NOT popping up when a user tries to double register. 

Stock purchasing form is in the works. Issues with how to present prices when a user types in a the ticker symbol. Currently there is a required reload button to see the price. Dynamic pricing while the user types the ticker symbol would be better, but not enough time.

Next will be to finish the stock purchase and selling form. Few other things to fix up as extra, but the stock form must be finished by tomorrow, Aug 5th.