Autocomplete Ticker lookup WORKS!!!
Stock price is dynamically updated while the ticker is being typed. Even when there are multiple stocks with the same sub-string, the stock with the highest current volume is presented.

Only thing left is to set up the buying and selling buttons, reset the database, and test, test, test, debug, test, test, debug, debug, test, tes.....